+++ include/lipm_walking/Sole.h
@@ -27,6 +27,8 @@
 
 #pragma once
 
+#include <mc_rtc/Configuration.h>
+
 namespace lipm_walking
 {
   /** Foot sole properties.
@@ -34,9 +36,12 @@ namespace lipm_walking
    */
   struct Sole
   {
-    double friction = 0.7;
-    double halfLength = 0.112; // [m]
-    double halfWidth = 0.065; // [m]
+    EIGEN_MAKE_ALIGNED_OPERATOR_NEW
+
+    Eigen::Vector2d leftAnkleOffset = Eigen::Vector2d::Zero(); /**< Offset from center to ankle frames in the left sole frame */
+    double friction = 0.7; /**< Friction coefficient */
+    double halfLength = 0.112; /**< Half-length of the sole in [m] */
+    double halfWidth = 0.065; /**< Half-width of the sole in [m]*/
   };
 }